+++ lib/dictosaurus.dart
@@ -63,7 +63,7 @@ export 'src/dictionary/_index.dart'
         TermVariantBase;
 export 'src/dictosaurus/_index.dart'
     show DictoSaurus, DictoSaurusMixin, DictoSaurusBase;
-export 'src/_common/_index.dart' show TermProperty, PartOfSpeech;
+export 'src/_common/_index.dart' show DictionaryEndpoint, PartOfSpeech;
 export 'package:gmconsult_dart_core/dart_core.dart' show Language;
 export 'package:gmconsult_dart_core/extensions.dart'
     show LanguageStringExtensions;